Refactor: 날짜선택 헤더UI 리팩터링 및 폴더구조 개선
