+++ src/pages/Booking/ui/SelectDateCarousel/SelectDateCarousel.type.ts
@@ -1,0 +1,9 @@
+export interface IArrowButtonProps {
+  onClick?: React.MouseEventHandler<HTMLButtonElement>
+}
+
+export interface IDateButtonProps {
+  date: string
+  today: string
+  onSelect: (date: string) => void
+}
