--- src/pages/Booking/ui/ui.type.ts
@@ -1,3 +1,0 @@
-export interface IArrowButtonProps {
-  onClick?: React.MouseEventHandler<HTMLButtonElement>
-}
